adding redis and kafka support for future and changing City's ID from Long's to Integer's because Cassandra's int type maps to 32-bit signed integers like Java's int so it is a better fit (also it was causing issues when searching for cities)

+++ hotel-service/src/main/java/io/hotely/hotel/repositories/CityRepository.java
@@ -8,4 +8,4 @@ import io.hotely.hotel.entities.City;
 
 @Transactional
 @Repository
-public interface CityRepository extends CassandraRepository<City, Long> {}
+public interface CityRepository extends CassandraRepository<City, Integer> {}
